+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -3,11 +3,11 @@ package kitchenpos.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kitchenpos.dao.MenuGroupRepository;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.dto.menuGroup.MenuGroupCreateRequest;
 import kitchenpos.dto.menuGroup.MenuGroupCreateResponse;
 import kitchenpos.dto.menuGroup.MenuGroupFindAllResponses;
+import kitchenpos.repository.MenuGroupRepository;
 
 @Service
 public class MenuGroupService {
