+++ src/main/java/kitchenpos/repository/OrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.repository;
 
 import java.util.List;
 import java.util.Optional;
